fix: fix issue with updating BPM would remove classes
Updating the `:bpm` assign in the socket was removing the
`bg-orange` class from the pads because the template was being
re-rendered. This was happening because we were not relying on the
`:sequence` assign in order to determine which pads should be
orange and which ones should be gray. This commit fixes that issue
by:

- Updating `SeqexWeb.LiveSequencer.render/1` to rely on the
  `:sequence` assign in order to determine the background color
  for each pad.
- Removing the `phx-hook` and `phx-update` attributes from the `<h1>`
  tag in the `SeqexWeb.LiveSequencer.render/1` function. We no longer
  need the Javascript functionality in order to toggle the background
  color for the pads.